renaming settingsContext to cap SettingsContext
